Don't overwrite reboot_cmd and fastsync when a caller specifies them
Our Chromium OS specific Host class uses our own reboot_cmd for speeding
up the reboot time. However, these arguments, reboot_cmd and fastsync,
force to be overwritten. As a result, the caller fails to specify its
own customized command.

This CL checks the arguments first and doesn't overwrite them if they
are specified by the caller.

BUG=chromium-os:33067;chrome-os-partner:12520
TEST=manual
Ran the original platform_BootDevice passed.
Changed platform_BootDevice reboot method to:
    self.client.reboot(reboot_cmd='sleep 10; reboot')
Ran and the log showed the above reboot command:
12:14:25 DEBUG| Running (ssh) 'sleep 10; reboot'

Reviewed-on: https://gerrit.chromium.org/gerrit/29976
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Ready: David James <davidjames@chromium.org>

+++ server/hosts/site_host.py
@@ -211,10 +211,12 @@ class SiteHost(remote.RemoteHost):
         sync should be finished in a short time during the reboot
         command.
         """
-        dargs['reboot_cmd'] = ('((reboot & sleep 10; reboot -f &)'
-                               ' </dev/null >/dev/null 2>&1 &)')
+        if 'reboot_cmd' not in dargs:
+            dargs['reboot_cmd'] = ('((reboot & sleep 10; reboot -f &)'
+                                   ' </dev/null >/dev/null 2>&1 &)')
         # Enable fastsync to avoid running extra sync commands before reboot.
-        dargs['fastsync'] = True
+        if 'fastsync' not in dargs:
+            dargs['fastsync'] = True
         super(SiteHost, self).reboot(**dargs)
